+++ app/src/main/java/nsapp/com/combienjtedois/model/Person.java
@@ -4,14 +4,14 @@ import java.io.Serializable;
 
 public class Person implements Serializable {
 
-    private final long id;
+    private final int id;
     private final String phoneNumber;
     private final String name;
     private final String modificationDate;
     private final String totalAmount;
     private String imageProfileUrl;
 
-    public Person(long id, String name, String totalAmount, String phoneNumber, String imageProfileUrl, String modificationDate) {
+    public Person(int id, String name, String totalAmount, String phoneNumber, String imageProfileUrl, String modificationDate) {
         this.id = id;
         this.name = name;
         this.modificationDate = modificationDate;
@@ -20,7 +20,7 @@ public class Person implements Serializable {
         this.imageProfileUrl = imageProfileUrl;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
